+++ resources/js/tools/PostScripts.js
@@ -1,35 +1,67 @@
+const delMsg = `
+  Attention, la suppression d'un article est definitive.
+  \n
+  Vous pouvez mettre en prive un article sans pour autant le supprimer.
+  \n
+  Etes-vous certains de vouloir le supprimer ?
+`
+
 export async function deletePost(id) {
-  const delMsg = `
-        Attention, la suppression d'un article est definitive.
-        \n
-        Vous pouvez mettre en prive un article sans pour autant le supprimer.
-        \n
-        Etes-vous certains de vouloir le supprimer ?
-    `
-  if (confirm(delMsg) === true) {
-    await axios({
-      method: 'delete',
-      url: `/blog/${id}/delete`,
-      withCredentials: true,
-    })
-      .then((res) => {
-        alert(res.data)
-        setTimeout(function () {
-          window.location.href = '/blog'
-        }, 1500)
+  let delAuth = null
+
+  await axios({
+    method: 'get',
+    url: `/blog/${id}/before-delete`,
+    withCredentials: true,
+  })
+    .then((res) => (delAuth = res.data.authorized))
+    .catch((err) => console.error(err))
+
+  if (delAuth === true) {
+    if (confirm(delMsg) === true) {
+      await axios({
+        method: 'delete',
+        url: `/blog/${id}/delete`,
+        withCredentials: true,
       })
-      .catch((err) => console.error(err))
-  } else {
-    alert("L'article ne sera pas supprime")
-  }
+        .then((res) => {
+          alert(res.data)
+          setTimeout(function () {
+            window.location.href = '/blog'
+          }, 1500)
+        })
+        .catch((err) => console.error(err))
+    } else {
+      alert("L'article ne sera pas supprime")
+    }
+  } else alert("Cet article n'est pas le votre")
 }
 
 export async function newStatus(id) {
+  let response = {}
+
   await axios({
     method: 'patch',
     url: `/blog/${id}/status`,
     withCredentials: true,
-  }).catch((err) => console.error(err))
+  })
+    .then(() => {
+      response = {
+        success: true,
+        msg: `L'etat a ete change`,
+      }
+    })
+    .catch((err) => {
+      console.error(err)
+      alert(err.response.data)
+
+      response = {
+        success: false,
+        msg: `Acces non authorise`,
+      }
+    })
+
+  return response
 }
 
 export function resetPostErrors(errors) {
